requirements.txt
python==3.8.18
joblib==1.3.2
numpy==1.24.4
packaging==23.2
paho-mqtt==1.6.1
pandas==2.0.3
patsy==0.5.6
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
statsmodels==0.14.1
threadpoolctl==3.2.0
tzdata==2023.4
